% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beset_elnet.R
\name{beset_elnet}
\alias{beset_elnet}
\title{Beset GLM with Elasticnet Regularization}
\usage{
beset_elnet(form, data, test_data = NULL, family = "gaussian",
  alpha = c(0.01, 0.5, 0.99), ..., n_folds = 5, n_repeats = 5,
  seed = 42, n_cores = 2)
}
\arguments{
\item{form}{A model \code{\link[stats]{formula}}.}

\item{data}{Either a \code{\link{data_partition}} object containing data sets
to be used for both model training and testing, or a single data frame that
will be used for model training only.}

\item{test_data}{Optional data frame with the variables in \code{form} and
the data to be used for model validation. Not necessary and ignored if a
\code{\link{data_partition}} object is passed via the \code{data} parameter.}

\item{family}{Character string naming the error distribution to be used in
the model. Currently supported options are \code{"gaussian"} (default),
\code{"binomial"}, and \code{"poisson"}.}

\item{alpha}{Numeric vector of alpha values between 0 and 1 to use as tuning
parameters. \code{alpha = 0} results in ridge regression, and \code{alpha =
1} results in lasso regression. Values in between result in a mixture of L1
and L2 penalties. (Values closer to 0 weight the L2 penalty more heavily,
and values closer to 1 weight the L1 penalty more heavily.)}

\item{n_folds}{Integer indicating the number of cross-validation folds.}

\item{n_repeats}{Number of times cross-validation should be repeated.}

\item{seed}{Seed for random number generator, used for assigning observations
to cross-validation folds.}

\item{n_cores}{Number of cores to use for parallel execution. If not
specified, the number of cores is set to 2. To determine the theoretical
maximum number of cores you have available, see
\code{\link[parallel]{detectCores}}, but note that the actual number of cores
available may be less. See \code{\link[parallel]{parallel}} for
more information.}
}
\description{
\code{beset_elnet} is a wrapper to \code{\link[glmnet]{glmnet}} for fitting
 generalized linear models via penalized maximum likelihood, providing
 automated data preprocessing and selection of both the elasticnet penalty
 and regularization parameter through repeated k-fold cross-validation.
}
\seealso{
\code{\link[glmnet]{glmnet}}
}
