% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beset_glmnet.R
\name{beset_glmnet}
\alias{beset_glmnet}
\title{Beset GLM with Elasticnet Regularization}
\usage{
beset_glmnet(form, train_data, test_data = NULL, alpha = c(0.05, 0.5, 0.95),
  cv_control = c(5, 5), impute = "none", seed = 42, n_cores = NULL)
}
\arguments{
\item{form}{A formula of the form y ~ x1 + x2 + ...}

\item{train_data}{Data frame containing the training data set and all of the
variables specified in \code{form}.}

\item{test_data}{Optional data frame containing an independent test data set
and all of the variables specified in \code{form}.}

\item{alpha}{Numeric vector of alpha values between 0 and 1 to use as tuning
parameters. \code{alpha = 0} results in ridge regression, and \code{alpha =
1} results in lasso regression. Values in between result in a mixture of L1
and L2 penalties. (Values closer to 0 weight the L2 penalty more heavily,
and values closer to 1 weight the L1 penalty more heavily.)}

\item{cv_control}{Integer vector of length two indicating the number of
cross-validation folds followed by the number of times cross-validation
should be repeated. Default is \code{c(5,5)}, i.e., 5-fold cross-validation,
repeated 5 times.}

\item{impute}{Optional character string giving a method for imputing missing
values. This must be one of the strings "none" (the default, which will
result in listwise deletion of any cases with missing values), "knn"
(K-nearest neighbors), "bag" (bagged trees), or "median".}

\item{seed}{Seed for random number generator.}

\item{n_cores}{Number of cores to use for parallel execution. If not
specified, the number of cores is set to the value of
\code{options("cores")}, if specified, or to approximately half the number of
cores detected by the \code{parallel} package. To determine the theoretical
maximum number of cores you have available, see
\code{\link[parallel]{detectCores}}, but note that the actual number of cores
available may be less. See \href{
https://stat.ethz.ch/R-manual/R-devel/library/parallel/doc/parallel.pdf}{
\code{parallel} package vignette} for more information.}
}
\description{
\code{beset_glmnet} is a wrapper to \code{\link[glmnet]{glmnet}} via
\code{\link[caret]{train}} for fitting generalized linear modesl via
penalized maximum likelihood, providing automated data preprocessing and
selection of both the elasticnet penalty and regularization parameter through
cross-validation.
}
\details{
\code{beset_glmnet} provides automatic centering and scaling of the predictor
variables (which is a must for regularized regression) and optional multiple
imputation of missing values. Imputation is repeated within each
cross-validation fold and thus variance among imputations becomes a part of
the cross-validation error.
}
\seealso{
\code{\link[glmnet]{glmnet}}, \code{\link[caret]{train}}
}

