% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beset_lm.R
\name{beset_lm}
\alias{beset_lm}
\title{Best Subset Selection for Least-Squares Regression}
\usage{
beset_lm(form, train_data, test_data = NULL, n_folds = 5, n_repeats = 5,
  seed = 42)
}
\arguments{
\item{form}{A model \code{\link[stats]{formula}}.}

\item{train_data}{A \code{\link[base]{data.frame}} with the variables in
\code{form} and the data to be used in model training.}

\item{test_data}{Optional \code{\link[base]{data.frame}} with the variables
in \code{form} and the data to be used in model testing.}

\item{seed}{An integer used to seed the random number generator when
assigning observations to folds.}

\item{k}{Single integer or integer vector of number of \code{k} folds to use
for cross-validation. By default, both 5 and 10 folds will be performed.}
}
\value{
A list with the following three components:
\enumerate{
 \item\describe{
   \item{best_model}{an object of class \code{\link[stats]{lm}} corresponding
   to the best linear model with the number of parameters with the smallest
   cross-validation error (largest cross-validation R-squared)}
   }
 \item\describe{
   \item{best_model_1SE}{an object of class \code{\link[stats]{lm}}
   corresponding to the best linear model with the smallest number of
   parameters within one standard error of the smallest cross-validation
   error (largest cross-validation R-squared)}
   }
 \item \code{\link[base]{data.frame}} with the following columns:
   \describe{
   \item{k_folds}{the number of k-folds used}
   \item{n_preds}{the number of predictors in model}
   \item{R2_train}{the mean in-fold (training) R-squared for each size of
     best model}
   \item{R2_train_SE}{the standard error of the mean in-fold (training)
     R-squared for each size of best model}
   \item{R2_cv}{the mean out-of-fold (test) R-squared for each size of best
     model}
   \item{R2_cv_SE}{the standard error of the mean out-of-fold (test)
     R-squared for each size of best model}
   \item{R2_test}{if \code{test_data} is provided, the mean R-squared when
     the model fit to each subsample of the training data is applied to the
     independent test set}
   \item{R2_test_SE}{if \code{test_data} is provided, the standard error of
     the mean R-squared when the model fit to each subsample of the training
     data is applied to the independent test set}
   }
 }
}
\description{
\code{beset_lm} performs best subset selection using cross-validation to find
 the optimal number of linear regression parameters.
}
\details{
\code{beset_lm} performs best subset selection, using
\code{\link[leaps]{regsubsets}} to fit a separate least-squares regression
for each possible combination of predictors, i.e., all models that contain
exactly 1 predictor, all models that contain exactly 2 predictors, and so
forth. For each number of predictors, \code{\link[leaps]{regsubsets}} picks
the best model, i.e., the one with the smallest residual sum of squares (RSS)
or equivalently the largest \eqn{R^2}{R-squared}. This results in a best fit
for every possible number of predictors. \code{beset_lm} then uses
\code{k}-fold cross-validation to select the "best of the best": the best
model with the number of predictors that minimizes the prediction RSS or
equivalently maximizes the prediction \eqn{R^2}{R-squared}, i.e., how well
the best models trained using \eqn{k - 1} folds predict the out-of-fold
sample. \code{beset_lm} uses \code{\link[caret]{createFolds}} to randomly
assign observations to \code{k} folds within subgroups based on percentiles
of the numeric outcome. This insures that the folds will be matched in terms
of the outcome's frequency distribution. \code{beset_lm} also insures the
reproducibility of your analysis by requiring a \code{seed} to the random
number generator as one of its arguments.
}
\section{Warnings}{

\enumerate{
 \item \code{beset_lm} handles missing data by performing listwise deletion.
  No other options for handling missing data are provided. The user is
  encouraged to deal with missing values prior to running this function.
 \item \code{beset_lm} is intended for use with additive linear models only.
   While there is no prohibition against the inclusion of interaction or
   polynomial terms, this practice is strongly discouraged. At best, this
   will result in an inefficient search because
   \code{\link[leaps]{regsubsets}} performs an exhaustive search over all
   possible variable subsets, including subsets that are hierarchically
   incomplete, i.e. subsets that contain an interaction term but are missing
   one or more of the subterms that comprise it. At worst, it may return one
   of these hierarchically incomplete models as the best model, an
   undesirable result if one cares about interpretability. If one wishes the
   model search to include interaction and/or non-linear effects, the
   \href{https://cran.r-project.org/web/packages/earth/index.html}{MARS}
   technique is recommended instead.
 \item \code{beset_lm} is best suited for searching over a small number of
   predictors (less than 10). For a large number of predictors (more than
   20), \code{\link{beset_elnet}} is recommended instead.
}
}
\seealso{
\code{\link[caret]{createFolds}}, \code{\link[leaps]{regsubsets}},
\code{\link[base]{set.seed}}, \code{\link{predict_R2}}
}

