% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beset_rf.R
\name{beset_rf}
\alias{beset_rf}
\alias{plot.beset_rf}
\title{Beset Random Forest}
\usage{
beset_rf(form, data, n_trees = 500, sample_rate = 0.632, mtry = NULL,
  min_obs_in_node = NULL, n_folds = 10, n_reps = 10, seed = 42,
  class_wt = NULL, cutoff = NULL, strata = NULL, parallel_type = NULL,
  n_cores = NULL, cl = NULL)

\method{plot}{beset_rf}(x, metric = c("auto", "mse", "rsq"), ...)
}
\description{
\code{beset_rf} is a wrapper to \code{\link[randomForest]{randomForest}} that
estimates predictive performance of the random forest using repeated k-fold
cross-validation. \code{beset_rf} insures that the correct arguments are
provided to \code{\link[randomForest]{randomForest}} and that enough
information is retained for compatibility with \code{beset} methods such as
variable \code{\link{importance}} and partial \code{\link{dependence}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: Plot method for "beset_rf" objects
}}

