% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_ci.R
\name{boot_ci}
\alias{boot_ci}
\alias{boot_cor}
\alias{boot_mean}
\title{Bootstrapped Confidence Intervals}
\usage{
boot_ci(data, statistic, n_rep = 1000, conf = 0.95, ...)

boot_mean(data, n_rep = 1000, conf = 0.95)

boot_cor(data, n_rep = 1000, conf = 0.95, ...)
}
\arguments{
\item{data}{The data as a vector, matrix or data frame. If it is a matrix or
data frame then each row is considered as one multivariate observation.}

\item{statistic}{A function which, when applied to data, returns a vector
containing the statistic(s) of interest. \code{statistic} whose first two
arguments must be as follows: the first argument must be the original data,
and the second argument must be a vector of indices, which the function must
reference when calculating the statistic. Any additional function arguments
must be assigned default values.}

\item{n_rep}{The number of bootstrap replicates.}

\item{conf}{A scalar or vector containing the confidence level(s) of the
required interval(s).}

\item{...}{Other named arguments for \code{statistic} which are passed
unchanged each time it is called.  For \code{boot_cor}, this could include
a \code{use} and/or \code{method} argument to override the defaults
documented in \code{\link{cor_list}}.}
}
\value{
Object of class "bootstrap", a list containing an object of class
"boot" and a stats data frame with the following columns:
\describe{
 \item{statistic}{observed statistic for each measurement}
 \item{ci_lower}{lower endpoint of the confidence interval}
 \item{ci_upper}{upper endpoint of the confidence interval}
}
}
\description{
Functions for calculating the bias corrected accelerated confidence intervals
from either univariate or multivariate data.
}
\section{Functions}{
\itemize{
\item \code{boot_mean}: Bootstrapped confidence interval for the mean

\item \code{boot_cor}: Bootstrapped confidence interval for Pearson's \emph{r}
}}
\seealso{
\code{\link[boot]{boot}}, \code{\link[boot]{boot.ci}},
 \code{\link{cor_loo}}, \code{\link{permute}}
}

