% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{cor}
\alias{cor}
\title{Correlation}
\usage{
cor(x, y = NULL, as_matrix = FALSE, use = "pairwise.complete.obs",
  method = "pearson")
}
\arguments{
\item{x}{a numeric vector, matrix or data frame.}

\item{y}{\code{NULL} (default) or a vector, matrix or data frame with
    compatible dimensions to \code{x}.   The default is equivalent to
    \code{y = x} (but more efficient).}

\item{as_matrix}{Logical indicating whether correlations should be returned
as a matrix.}

\item{use}{an optional character string giving a
    method for computing covariances in the presence
    of missing values.  This must be (an abbreviation of) one of the strings
    \code{"everything"}, \code{"all.obs"}, \code{"complete.obs"},
    \code{"na.or.complete"}, or \code{"pairwise.complete.obs"}.}

\item{method}{a character string indicating which correlation
    coefficient (or covariance) is to be computed.  One of
    \code{"pearson"} (default), \code{"kendall"}, or \code{"spearman"}:
    can be abbreviated.}
}
\value{
If \code{as_matrix = FALSE} (the default), a \code{\link{cor_list}}
object. If \code{as_matrix = TRUE}, a correlation matrix. Both are assigned
the attribute "coef" which indicates the statistic that is being returned,
e.g., Pearson's \emph{r}, Kendall's \emph{tau}, or Spearman's \emph{rho}.
}
\description{
This function overrides the \code{\link[stats]{cor}} function provided by the
stats library to give the option of returning correlations as a
\code{\link{cor_list}} object.
}
\details{
\code{cor} differs from R's native \code{\link[stats]{cor}} only in that it
uses pairwise-complete observations and returns a \code{\link{cor_list}}
object by default. If you want \code{beset::cor} to mimic the default
behavior of \code{stats::cor}, set \code{as_matrix = TRUE} and
\code{use = "everything"}.
}
\seealso{
\code{\link[stats]{cor}}, \code{\link{cor_list}}
}

