% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_dif.R
\name{cor_dif}
\alias{cor_dif}
\title{Robust Significance Testing for Between-Group Differences in Correlations}
\usage{
cor_dif(form, data, n_iter = 1000, loo = TRUE)
}
\arguments{
\item{form}{A formula of the form \code{groups ~ x1 + x2 + ...}. That is, the
left-hand side indicates the grouping factor and the right-hand side
indicates the (numeric) measurements to correlate within each group.}

\item{data}{Data frame from which variables specified in \code{form} are to
be taken.}

\item{n_iter}{Number of iterations to use for the bootstrap and permutation
tests.}

\item{loo}{Logical indicating whether permutation tests for group differences
should be performed on the original correlation coefficients (\code{FALSE})
or on the leave-one-out coefficients (\code{TRUE}).}
}
\description{
Given the same set of variables from two independent samples, \code{cor_dif}
calculates bivariate correlations for each sample and tests whether each
correlation is signficantly different between samples.
}
\details{
\code{cor_dif} was designed to test multiple hypotheses of group differences
in correlations between elements of a network, such as whether a treatment
alters the correlation between the activity of two brain regions. Fisher's z
test is traditionally used for this purpose, but may yield inaccurate
probability estiamtes when a) one or both variables being correlated is not
normally distributed, b) the sample size for each group is small (n <= 10),
or c) there are multiple pairwise correlations from the same data set.
Although the latter is a multiple-comparisons problem, conventional
statistical corrections assume that the hypothesis tests are independent,
which is assuredly not the case for multiple hypothesis tests on correlations
arising from a networked system.

\code{cor_dif} can be used with small sample sizes and, by default, uses
leave-one-out cross validation to produce more realistic estimates of the
correlation coefficients themselves, provides bootstrapped estimates of the
95\% confidence intervals for the correlation coefficients, and conducts
permutation tests to calculate p-values conditioned on the number of
correlations being assessed. (For example, if 15 correlation coefficients are
being computed, then the resulting p value corresponds to the probability of
calulcating 15 correlation differences and obtaining at least one as extreme
as the observed difference.)
}
\seealso{
\code{\link{cor_loo}}, \code{\link{boot_cor}},
\code{\link{permute_cor_dif}}
}

