% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_loo.R
\name{cor_loo}
\alias{cor_loo}
\title{Leave-One-Out Correlations}
\usage{
cor_loo(data)
}
\arguments{
\item{data}{A numeric matrix or data frame containing only numeric variables.}
}
\value{
A numeric vector of correlation coefficients, named according to each
 unique pairwise combination of variables.
}
\description{
\code{cor_loo} computes the leave-one-out estimate for all Pearson
correlations among the columns of a matrix or data frame.
}
\details{
The leave-one-out (LOO) cross-validation correlation provides an estimate of
Pearson's \eqn{r} that corrects for inflation in the absolute magnitutde of
\eqn{r} associated with small sample sizes, in which single obvervations can
have undue influence on the size of the correlation coefficient. It is
equivalent to fitting a linear model to all possible sets of \eqn{n-1}
observations, predicting the left-out observation, and using the root mean
square error (RMSE) of the predicitions to recalculate \eqn{r}. With linear
models, this quantity can be computed without repeating the leave-one-out
procedure \eqn{N} times by using the companion \code{\link{predict_R2}}
function, which computes the predicted \eqn{r^2}{r-squared} using the PRESS
statistic. The LOO estimate of \eqn{r} is then given as the square root of
the predicted \eqn{r^2}{r-squared} for each bivariate linear regression
model, with the sign of the LOO \eqn{r} assigned to match that of the
corresponding \eqn{beta} coefficient.
}
\section{Note}{

Correlations are computed using pairwise complete observations.
}
\seealso{
\code{\link{predict_R2}}
}

