% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folds.R
\name{create_folds}
\alias{create_folds}
\title{Create Folds for Cross-Validation}
\usage{
create_folds(y, k = 10, reps = 1, seed = 42)
}
\arguments{
\item{y}{A vector of outcomes.}

\item{k}{An integer for the number of folds (default = 10).}

\item{reps}{An integer for the number of repetitions (default = 1).}

\item{seed}{An integer used to reproduce the random fold assignments.}
}
\value{
A list of \eqn{k * reps} row-position integers corresponding to the
training data for each independent fold of each repetition. (Test data are
implied by the rows that are omitted.) The names of the list elements give
the fold membership using the pattern "Fold\emph{i}.Rep\emph{j}".
}
\description{
\code{create_folds} randomly splits the data into \code{k} folds, with
optional repetition of the splitting procedure under different randomization
seeds
}
\details{
\code{create_folds} is used by \code{beset} modeling functions to assign
folds, stratified by the response variable (\code{y}). When \code{y} is a
factor, random fold assignment is done within the levels of \code{y} to
balance the class distribution across folds. When \code{y} is numeric, the
sample is split into strata based on percentiles and random fold assignment
is done within strata. The number of groups is set dynamically based on
sample size and \code{k}. In addition, if \code{y} is zero-inflated, the
algorithm attempts to balance the proportion of \code{0} values across folds.
Optionally, if the \code{reps} argument is set to a value greater than
\code{1}, the function repeats the stratified fold assignment by this number
of \code{reps}, using different random seeds each time.
}
\section{Note}{

Previous versions of \code{beset} imported the function
\code{\link[caret]{createMultiFolds}} to set up repeated k-fold
cross-validation. While the object returned by \code{create_folds} copies the
structure of what would be returned by \code{\link[caret]{createMultiFolds}},
the algorithm for fold assignment is different. This may result in slightly
different cross-validation results if you attempt to reproduce an analysis
performed with a previous version.
}

\seealso{
\code{\link{data_partition}}
}
