% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_R2.R
\name{cv_r2}
\alias{cv_r2}
\title{Cross-Validated R-Squared}
\usage{
cv_r2(object, n_cores = 2, n_folds = 10, n_repeats = 10, seed = 42)
}
\arguments{
\item{n_cores}{Integer value indicating the number of workers to run in
parallel during subset search and cross-validation. By default, this will
be set to 2. To determine the theoretical maximum number of cores you have
available, see \code{\link[parallel]{detectCores}}, but note that the actual
number of cores available may be less. See
\code{\link[parallel]{parallel-package}} for more information.}

\item{n_folds}{Integer indicating the number of folds to use for
cross-validation.}

\item{n_repeats}{Integer indicating the number of times cross-validation
should be repeated.}

\item{seed}{An integer used to seed the random number generator when
assigning observations to folds.}
}
\description{
Returns a cross-validated estimate of R-squared, or how well the fitted model
object is expected to explain new data.
}
