% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{dependence}
\alias{dependence}
\alias{dependence.beset_elnet}
\alias{dependence.nested_elnet}
\alias{dependence.randomForest}
\alias{dependence.beset_rf}
\alias{plot.part_depend}
\title{Partial-Dependence Effects}
\usage{
dependence(object, ...)

\method{dependence}{beset_elnet}(object, x = NULL, cond = NULL,
  alpha = NULL, lambda = NULL, metric = c("mce", "auc", "mae", "mce",
  "mse"), error = c("auto", "btwn_fold_se", "btwn_rep_range", "none"),
  x_lab = NULL, y_lab = NULL, ...)

\method{dependence}{nested_elnet}(object, x = NULL, cond = NULL,
  alpha = NULL, lambda = NULL, metric = c("mce", "auc", "mae", "mce",
  "mse"), error = c("auto", "btwn_fold_se", "btwn_rep_range", "none"),
  x_lab = NULL, y_lab = NULL, parallel_type = NULL, n_cores = NULL,
  cl = NULL, ...)

\method{dependence}{randomForest}(object, data = NULL, x = NULL, y = NULL,
  cond = NULL, x_lab = NULL, y_lab = NULL, make_plot = TRUE, ...)

\method{dependence}{beset_rf}(object, x = NULL, cond = NULL, x_lab = NULL,
  y_lab = NULL, ..., parallel_type = NULL, n_cores = NULL, cl = NULL)

\method{plot}{part_depend}(x, p_max = 16, ...)
}
\arguments{
\item{object}{Model or list of resampled models}

\item{x}{Character string giving name(s)}

\item{alpha}{(Optional) \code{numeric} value to force selection of
elastic-net model with the given \code{alpha} parameter. If left \code{NULL},
the best value of \code{alpha} will be chosen using the cross-validation
\code{metric} and \code{oneSE} rule.}

\item{lambda}{(Optional) \code{numeric} value to force selection of elastic-
net model with the given \code{lambda} parameter. If left \code{NULL},
the best value of \code{lambda} will be chosen using the cross-validation
\code{metric} and \code{oneSE} rule.}

\item{metric}{\code{Character} string giving prediction metric on which to
base model selection. Can be one of \code{"mce"} (mean cross entropy--the
default), \code{"mse"} (mean squared error), \code{"aic"} (Akaike informatio
criterion, not applicable for elastic net), \code{"auc"} (area under the ROC
curve--only applicable if response is binomial), or \code{"mae"} (mean
absolute error--only applicable if response is numeric).}
}
\description{
Generate data for partial dependence plots.
}
