% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best.R
\name{get_best}
\alias{get_best}
\alias{get_best.beset_glm}
\alias{get_best.beset_elnet}
\title{Get best model from a "beset" class object}
\usage{
get_best(object, ...)

\method{get_best}{beset_glm}(object, n_pred = NULL, metric, oneSE = TRUE)

\method{get_best}{beset_elnet}(object, alpha = NULL, lambda = NULL,
  metric = c("mce", "auc", "mae", "mce", "mse"), oneSE = TRUE)
}
\arguments{
\item{object}{A \code{\link{beset_glm}} or \code{\link{beset_elnet}} object.}

\item{n_pred}{(Optional) \code{integer} number of predictors that the best
model should contain. If specified, all other arguments are ignored.}

\item{metric}{\code{Character} string giving prediction metric on which to
base model selection. Can be one of \code{"mce"} (mean cross entropy--the
default), \code{"mse"} (mean squared error), \code{"aic"} (Akaike informatio
criterion, not applicable for elastic net), \code{"auc"} (area under the ROC
curve--only applicable if response is binomial), or \code{"mae"} (mean
absolute error--only applicable if response is numeric).}

\item{oneSE}{\code{Logical} indicating whether or not to use the "one
standard error" rule. If \code{TRUE} (default) the simplest model within one
standard error of the optimal model is returned. If \code{FALSE} the model
with the optimal cross-validation performance is returned.}

\item{alpha}{(Optional) \code{numeric} value to force selection of
elastic-net model with the given \code{alpha} parameter. If left \code{NULL},
the best value of \code{alpha} will be chosen using the cross-validation
\code{metric} and \code{oneSE} rule.}

\item{lambda}{(Optional) \code{numeric} value to force selection of elastic-
net model with the given \code{lambda} parameter. If left \code{NULL},
the best value of \code{lambda} will be chosen using the cross-validation
\code{metric} and \code{oneSE} rule.}
}
\description{
\code{get_best} is a generic function used to obtain the best performing
predictive model from \code{\link{beset}} objects based on information
or cross-validation metrics.
}
\section{Methods (by class)}{
\itemize{
\item \code{beset_glm}: Get best model from a "beset_glm" object

\item \code{beset_elnet}: Get best model from a "beset_elnet" object
}}

