% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance}
\alias{importance}
\alias{importance.nested}
\alias{importance.beset_rf}
\alias{plot.variable_importance}
\title{Relative Variable Importance}
\usage{
importance(object, ...)

\method{importance}{nested}(object, ...)

\method{importance}{beset_rf}(object, ...)

\method{plot}{variable_importance}(x, p_max = 20, labels = NULL)
}
\arguments{
\item{object}{A model object for which variable importance scores are
desired.}

\item{...}{Additional named arguments that define the model selection rules.
See \code{\link{summary.beset}}.}
}
\description{
\code{importance} is a generic function for obtaining resampled variable
importance scores. The function invokes particular
\code{\link[utils]{methods}} which depend on the \code{\link[base]{class}}
of the first argument.
}
\section{Methods (by class)}{
\itemize{
\item \code{nested}: Determine variable importance for "nested"
\code{\link{beset_glm}} and \code{\link{beset_elnet}} objects

\item \code{beset_rf}: Relative importance method for "beset_rf" objects

\item \code{variable_importance}: Plot method for "variable_importance" objects
}}

