% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.cross_valid.R
\name{mean.cross_valid}
\alias{mean.cross_valid}
\title{Cross-Validation Mean}
\usage{
\method{mean}{cross_valid}(x, ...)
}
\arguments{
\item{x}{A "cross_valid" object returned by \code{\link{validate}}}

\item{...}{Additional "cross_valid" objects.}
}
\description{
Estimate cross-validation performance that would be obtained by averaging
predictions of more than one model.
}
