% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition}
\alias{partition}
\title{Partition Data into Training and Test Sets}
\usage{
partition(data, y, p = 0.75, seed = 42)
}
\arguments{
\item{data}{Data frame to be partitioned.}

\item{y}{Name of the response variable, without quotes.}

\item{p}{The fraction of data that should be included in the training set.}

\item{seed}{A single value, interpreted as an integer, for seeding the random
number generator. See \code{\link[base]{set.seed}}.}
}
\value{
An object of class "data_partition": a list containing two data
frames named \code{train} and \code{test}, containing the training and
testing sets, respectively.
}
\description{
\code{partition} randomly splits a data frame into two data frames,
\code{train} and \code{test}, which are returned as a
\code{\link{data_partition}} structure.
}
\details{
\code{partition} creates a train/test split among the rows of a data frame
based on stratified random sampling within the factor levels of a
classification outcome or the quintiles of a numeric outcome. This insures
that the training and test samples will be closely matched in terms of class
incidence or frequency distribution of the outcome measure. \code{partition}
includes a \code{seed} argument so that the randomized partitioning is
reproducible. The \code{train} and \code{test} data frames are returned
bound together in a \code{\link{data_partition}} structure so that their
common ancestry is maintained and self-documented. For example, if you name
your \code{\link{data_partition}} "\code{data}", you can intuitively access
the training set with \code{data$train} and its corresponding test set with
\code{data$test}.
}
\seealso{
\code{\link[base]{set.seed}}, \code{\link{data_partition}}
}
