% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{permute}
\alias{permute}
\alias{permute_cor_dif}
\alias{permute_mean_dif}
\title{Permutation Test for Multiple Comparisons}
\usage{
permute(form, data, statistic, n_perm = 1000)

permute_mean_dif(form, data, n_perm = 1000)

permute_cor_dif(form, data, loo = TRUE, n_perm = 1000)
}
\arguments{
\item{form}{A formula of the form \code{group ~ var1 + var2 + ...}. That is,
the left-hand side specifies the grouping factor, and the right-hand side
specifies the (numeric) variables that are to be evaluated for group
differences.}

\item{data}{Data frame from which variables specified in \code{form} are to
be taken.}

\item{statistic}{A function which when applied to data returns a vector
containing the effect-size statistic(s) of interest. \code{statistic} must
take exactly two arguments. The first argument passed will be a vector of
group labels. The second will be a data frame of measurements whose rows
correspond to the group labels. \strong{If the statistic is signed (i.e., it
reflects the direction of the group difference) and you desire a two-tailed
test, be sure that \code{statistic} returns the \emph{absolute value} of the
effect-size statistic.}}

\item{n_perm}{The number of permutations to perform.}

\item{loo}{For \code{permute_cor_dif}, logical indicating whether to use
leave-one-out estimates of the correlation coefficients (\code{TRUE} by
default) or the original coefficients (\code{FALSE}).}
}
\value{
Object of class "permutation", a list containing the following items:
\describe{
 \item{sampling_distribution}{Numeric vector containing the maximum group
 difference in terms of absolute value (largest group difference regardless
 of direction) obtained for each random permutation of the group labels. This
 corresponds to the distribution of the maximum effect sizes that would be
 obtained from your multivariate data set under the null hypothesis.}
\item{effect_size}{Numeric vector containing the effect size for each of the
 dependent variables. For \code{permute_mean_dif} and \code{permute_cor_dif}
 this will be Cohen's \eqn{d}}
\item{p_value}{Numeric vector containing the empirical probability of
obtaining the observed test statistic for each of the dependent variables,
given that the null hypothesis is true and conditioned on the number of
variables assessed and their covariance.}
}
}
\description{
Computes the empirical distribution of the maximum of a test statistic
repeated across multiple measurements under the null hypothesis.
}
\section{Functions}{
\itemize{
\item \code{permute_mean_dif}: Permutation test for mean differences between groups.

\item \code{permute_cor_dif}: Permutation test for correlation differences between
groups. \code{permute_cor_dif} performs a permutation test on the
between-group z-score difference for all bivariate corrleations obtained
from two independent samples.
}}
\section{Warning}{

If you are writing your own \code{statistic} function, be sure that either
your variables are centered and scaled or, better yet, that your
\code{statistic} is itself standardized.
}
\seealso{
\code{\link{boot_ci}}, \code{\link{cor_dif}}
}

