% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.beset}
\alias{plot.beset}
\alias{plot.beset_elnet}
\alias{plot.beset_glm}
\alias{plot.beset_zeroinfl}
\title{Plot Methods for \code{beset} Objects}
\usage{
\method{plot}{beset_elnet}(x, type = "cv", metric = NULL, ...)

\method{plot}{beset_glm}(x, metric = NULL, se = TRUE, ...)

\method{plot}{beset_zeroinfl}(x, type = "cv", metric = "mce", se = TRUE,
  ...)
}
\arguments{
\item{x}{An object of class \code{"beset_glm"},
\code{"beset_zeroinfl"}, or \code{"beset_elnet"}}

\item{type}{Type of error to plot. Can be one of \code{"train"}, \code{"cv"},
or \code{"test"}.}

\item{metric}{Which error metric to plot. Can be one of \code{"dev"} for
deviance, \code{"mae"} for mean absolute error, \code{"mce"} for mean cross
entropy, \code{"mse"} for mean squared error, or \code{"r2"} for R-squared.}

\item{...}{Arguments to be passed to methods}
}
\description{
Plot Methods for \code{beset} Objects
}
