% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.beset_elnet.R
\name{plot.beset_elnet}
\alias{plot.beset_elnet}
\title{Plot Method for the \code{beset_elnet} Class}
\usage{
\method{plot}{beset_elnet}(object, title = "")
}
\arguments{
\item{object}{An object of class \code{beset_elnet}.}

\item{title}{Optional character vector giving plot title.}
}
\description{
\code{plot.beset_elnet} takes the output of a \code{\link{beset_elnet}}
 object and creates a line plot using \code{\link[ggplot2]{ggplot}}.
}
\details{
\code{plot.beset_elnet} produces a plot showing cross-validation error as a
function of alpha (L1-L2 mixing parameter) and lambda (regularization
parameter) from a cross-validated elastic net regression, as implemented by
 \code{\link{beset_elnet}}.
}
