% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.beset_lm.R
\name{plot.beset_lm}
\alias{plot.beset_lm}
\title{Plot Method for the \code{beset_lm} Class}
\usage{
\method{plot}{beset_lm}(object, SE = TRUE, title = "")
}
\arguments{
\item{object}{An object of class \code{beset_lm}.}

\item{SE}{Logical indicating whether or not standard-error bars should be
plotted.}

\item{title}{Optional character vector giving plot title.}
}
\description{
\code{plot.beset_lm} takes the output of a \code{\link{beset_lm}} object and creates
a line plot using \code{\link[ggplot2]{ggplot}}.
}
\details{
\code{plot.beset_lm} produces a plot showing \eqn{R^2} as a function of the
number of model parameters from a cross-validated best-subset-selection
procedure, as implemented by \code{\link{beset_lm}}. Line graphs are plotted
for the training, cross-validated, and, if included, independent-test
\eqn{R^2}, which are labeled in the legend as "Train", "CV", and "Test",
respectively. If more than one choice of \code{k} was used for cross-
validation, separate panels are graphed for each.
}

