% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.beset_zeroinfl.R
\name{plot.beset_zeroinfl}
\alias{plot.beset_zeroinfl}
\title{Plot Method for the \code{beset_zeroinfl} Class}
\usage{
\method{plot}{beset_zeroinfl}(object, metric = "MCE", se = TRUE)
}
\arguments{
\item{object}{An object of class \code{beset_zeroinfl}.}

\item{SE}{Logical indicating whether or not standard-error bars should be
plotted.}
}
\description{
\code{plot.beset_zeroinfl} takes the output of a \code{\link{beset_zeroinfl}} object and creates
a line plot using \code{\link[ggplot2]{ggplot}}.
}
\details{
\code{plot.beset_zeroinfl} produces a plot showing cross-validated prediction
error as a function of the number of model parameters from a
\code{\link{beset_zeroinfl}} object. Line graphs are plotted for the
training, cross-validated, and, if included, independent-test error, which
are labeled in the legend as "Train", "CV", and "Test", respectively.
}
