% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_R2.R
\name{predict_R2}
\alias{predict_R2}
\title{Predicted R-squared}
\usage{
predict_R2(object, newdata = NULL)
}
\arguments{
\item{object}{Object of class inheriting from "lm"}

\item{newdata}{An optional data frame in which to look for variables with
which to predict. If omitted, leave-one-out cross validation is used to
simulate predictions on new data.}
}
\value{
Predicted R-squared as 1 - RSS / TSS, where RSS corresponds to the
sum of the squared residuals (observed y - predicted y) and TSS is the total
sum of squares.
}
\description{
\code{predict_R2} calculates predicted R-squared, which indicates how well a
regression model predicts responses for new observations.
}
\details{
\code{predict_R2} provides a metric analagous to R-squared, except instead of
describing how well a linear model fits the original data, it describes how
well the model predicts new data. If new data are not provided then the
predicted R-square will be estimated using the PRESS statistic
(predicted residual sum of squares), which is given by the sum of squares of
the ratio of the model residuals to the leverages (the diagonal elements of
the hat matrix) subtracted from 1. This provides for fast computation of the
leave-one-out residuals from the data returned by a single call to
\code{\link[base]{lm}}, without needing to recompute the model N times after
removing each of N observations. Note that predicted R-squared can be negative
if overfitting is severe; i.e., when predicting new data, the model performs
worse than if one were to always predict the sample mean.
}

