% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2d2.R
\name{r2d2}
\alias{r2d2}
\title{R-squared as Deviance Explained}
\usage{
r2d2(x, ...)
}
\value{
R-squared statistic based on the deviance ratio.
}
\description{
\code{r2d2} calculates an R-squared defined as the proportionate reduction in
uncertainty, measured by Kullback-Leibler divergence, due to the fitted
model. This corresponds to the fraction of deviance explained, which, for
linear models with normally distributed residuals, is equivalent to the
traditional R-squared.
}
\details{
For standard linear regression models fit with \code{\link[stats]{lm}}, the
familiar coefficient of determination, R-squared, can be obtained with
\code{\link[stats]{summary.lm}}. However, R-squared is not provided for
generalized linear models (GLMs) fit with \code{\link[stats]{glm}},
presumably because it can have undesirable properties when applied to GLMs
if the error distribution is non-normal (e.g., binomial, Poisson, etc.).
For example, it can lie outside the [0,1] interval and decrease as predictors
are added. Most importantly, the interpretation of R-squared as the fraction
of uncertainty explained by the model does not generally hold for exponential
family regression models.

Cameron and Windmeijer (1997) proposed an R-squared measure (termed
\eqn{R_{KL}^2}) for GLMs based on Kullback-Leibler (KL) divergence
(\eqn{entropy -  cross-entropy}), which restores all of the desirable
properties of R-squared, including its interpretation as the fraction of
uncertainty explained. Owing to an equivalence between KL-divergence and
deviance, others have referred to this metric as \eqn{R_D^2} (Martin & Hall,
2016), \eqn{D^2} (\code{\link[modEvA]{Dsquared}}), or \code{dev.ratio}
(\code{\link[glmnet]{glmnet}}). It is defined as \eqn{1-dev/nulldev},
where \eqn{dev} is the deviance: \eqn{2*(loglik_sat - loglik_fit)}, where
\eqn{loglik_sat} is the log-likelihood for the saturated model (a model
witha free parameter per observation) and \eqn{loglik_fit} is the
log-likelihood for the fitted model; and where \eqn{nulldev} is the null
deviance: \eqn{2*(loglik_sat - loglik_null)}, where \eqn{loglik_null} is the
log-likelihood for the intercept-only model. Following the reasoning of
Martin & Hall (2016), the saturated and null models for ZI regression are
taken to be the equivalent saturated and null models for the corresponding
non-ZI regression: e.g., the saturated model for ZI-Poisson regression is the
same as the saturated model for Poisson regression.
}
\references{
Colin Cameron A, Windmeijer FAG. (1997) An R-squared measure of goodness
of fit for some common nonlinear regression models. J Econometrics, 77,
329–342.

Jacob Martin & Daniel B. Hall (2016) R2 measures for zero-inflated regression
models for count data with excess zeros, Journal of Statistical Computation
and Simulation, 86:18, 3777-3790, DOI: 10.1080/00949655.2016.1186166
}
\seealso{
\code{\link{prediction_metrics}}, \code{\link{deviance.zeroinfl}}
}
