% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.beset_glm.R
\name{summary.beset_glm}
\alias{summary.beset_glm}
\title{Summary Method for the \code{beset_glm} Class}
\usage{
\method{summary}{beset_glm}(object, metric = "MCE", n_pred = NULL,
  oneSE = TRUE, n_cores = 2)
}
\arguments{
\item{object}{An object of class \code{\link{beset_glm}}.}
}
\description{
\code{summary.beset_glm} summarizes the output of a \code{\link{beset_glm}}
 object.
}
