% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.beset_lm.R
\name{summary.beset_lm}
\alias{summary.beset_lm}
\title{Summary Method for the \code{beset_lm} Class}
\usage{
\method{summary}{beset_lm}(object)
}
\arguments{
\item{object}{An object of class \code{\link{beset_lm}}.}
}
\description{
\code{summary.beset_lm} summarizes the output of a \code{\link{beset_lm}}
 object.
}

