% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.beset_zeroinfl.R
\name{summary.beset_zeroinfl}
\alias{summary.beset_zeroinfl}
\title{Summary Method for the \code{beset_zeroinfl} Class}
\usage{
\method{summary}{beset_zeroinfl}(object, n_pred = NULL, metric = "MCE",
  oneSE = TRUE, n_cores = 2)
}
\arguments{
\item{object}{An object of class \code{\link{beset_zeroinfl}}.}
}
\description{
\code{summary.beset_zeroinfl} summarizes the output of a
\code{\link{beset_zeroinfl}} object.
}
