% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_methods.R
\name{summary.bootstrap}
\alias{summary.bootstrap}
\title{Summarizing Bootstrap Objects}
\usage{
\method{summary}{bootstrap}(object, ...)
}
\arguments{
\item{object}{An object of class \code{bootstrap}.}
}
\description{
Methods for the \code{bootstrap} class.
}

