% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\alias{validate.lm}
\alias{validate.glm}
\alias{validate.zeroinfl}
\alias{validate.glmnet}
\alias{validate.beset}
\alias{validate.nested}
\alias{validate.randomForest}
\title{Cross-Validated Prediction Metrics}
\usage{
validate(object, ...)

\method{validate}{lm}(
  object,
  data = NULL,
  n_folds = 10,
  n_reps = 10,
  seed = 42,
  silent = FALSE,
  ...
)

\method{validate}{glm}(
  object,
  data = NULL,
  n_folds = 10,
  n_reps = 10,
  seed = 42,
  silent = FALSE,
  ...
)

\method{validate}{zeroinfl}(object, n_folds = 10, n_reps = 10, seed = 42, silent = FALSE, ...)

\method{validate}{glmnet}(
  object,
  x = NULL,
  y = NULL,
  lambda = NULL,
  offset = NULL,
  weights = NULL,
  n_folds = 10,
  n_reps = 10,
  seed = 42,
  envir = .GlobalEnv,
  silent = FALSE,
  ...
)

\method{validate}{beset}(object, ...)

\method{validate}{nested}(object, metric = "auto", oneSE = TRUE, ...)

\method{validate}{randomForest}(
  object,
  data = NULL,
  x = NULL,
  n_folds = 10,
  n_reps = 10,
  seed = 42,
  ...,
  parallel_type = NULL,
  n_cores = NULL,
  cl = NULL,
  silent = FALSE
)
}
\arguments{
\item{object}{A model object for which a cross-validated R-squared is
desired.}

\item{...}{Additional arguments passed to model summary methods.}

\item{data}{Data frame that was used to train the model. Only needed if
the training data is not contained in the model \code{object}.}

\item{n_folds}{\code{Integer} indicating the number of folds to use for
cross-validation.}

\item{n_reps}{\code{Integer} indicating the number of times cross-validation
should be repeated.}

\item{seed}{\code{Integer} used to seed the random number generator.}

\item{silent}{\code{Logical} indicating whether to suppress warning messsages
related to autocorrection of n_folds and n_reps parameters when they are
inappropriate for sample size or cross-validation method

object call}

\item{x}{Model matrix that was used to train elastic net.}

\item{y}{Response variable that was used to train elastic net.}

\item{lambda}{\code{Numeric} value of the penalty parameter \code{lambda}
at which predictions are cross-validated. Default is the entire sequence
used to create the model, but importance scores will only be generated
if a single \code{lambda} value is given.}

\item{offset}{\code{Numeric} offset vector that was used to train elastic
net.}

\item{weights}{= \code{Numeric} offset vector that was used to train
elastic net.}

\item{envir}{\code{Environment} in which to look for variables listed in}

\item{metric}{\code{Character} string giving prediction metric on which to
base model selection. Can be one of \code{"auc"} for area under the (ROC)
curve (only available for binomial family), \code{"mae"} for mean absolute
error (not available for binomial family), \code{"mae"} for mean absolute
error, \code{"mce"} for mean cross entropy, or \code{"mse"} for mean squared
error. Default is \code{"auto"} which plots MSE for Gaussian-family models
and MCE for all other families.}

\item{oneSE}{\code{Logical} indicating whether or not to use the "one
standard error" rule. If \code{TRUE} (default) the simplest model within one
standard error of the optimal model is returned. If \code{FALSE} the model
with the optimal cross-validation performance is returned.}

\item{parallel_type}{(Optional) character string indicating the type of
parallel operation to be used, either \code{"fork"} or \code{"sock"}. If
omitted and \code{n_cores > 1}, the default is \code{"sock"} for Windows and
otherwise either \code{"fork"} or \code{"sock"} depending on which process is
being run.}

\item{n_cores}{Integer value indicating the number of workers to run in
parallel during subset search and cross-validation. By default, this will
be set to one fewer than the maximum number of physical cores you have
available, as indicated by \code{\link[parallel]{detectCores}}. Set to 1 to
disable parallel processing.}

\item{cl}{(Optional) \code{\link[parallel]{parallel}} or
\code{\link[snow]{snow}} cluster for use if \code{parallel_type = "sock"}.
If not supplied, a cluster on the local machine is automatically created.}
}
\value{
a "\code{cross_valid}" object consisting of a list with the
following elements:
\describe{
  \item{cv_stats}{a list of cross-validated prediction metrics, each
  containing the mean, between-fold standard error ("btwn_fold_se"), and
  between-repetition min-max range ("btwn_rep_range") of each metric.}
  \item{predictions}{a data frame containing the
  hold-out predictions for each row in the training data, with a separate
  column for each repetition of the k-fold cross-validation}
  \item{fold_assignments}{a data frame of equal
  dimensions to \code{predictions} giving the number of the hold-out fold
  of the corresponding element in \code{predictions}}
  \item{parameters}{a list documenting how many folds and repetitions were used
  for cross-validation, and the seed passed to the random number generator,
  which will be needed to reproduce the random fold assignments}
  }
}
\description{
\code{validate} is a generic function for cross-validating predictions from
the results of various model fitting functions. The function invokes
particular \code{\link[utils]{methods}} which depend on the
\code{\link[base]{class}} of the first argument.
}
\details{
To obtain cross-validation statistics, first fit a model as you
normally would and then pass the model object to \code{validate}.
}
\section{Methods (by class)}{
\itemize{
\item \code{validate(lm)}: Cross-validation of linear models

\item \code{validate(glm)}: Cross-validation of GLMs

\item \code{validate(zeroinfl)}: Cross-validation of GLMs

\item \code{validate(glmnet)}: Cross-validation of GLM nets

\item \code{validate(beset)}: Cross-validation of beset objects

\item \code{validate(nested)}: Extract test error estimates from "nested beset"
objects with nested cross-validation

\item \code{validate(randomForest)}: Cross-validation of random forests

}}
