% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beset.R
\docType{package}
\name{beset}
\alias{beset}
\alias{beset-package}
\title{beset: Best Subset Predictive Modeling}
\description{
\code{beset} is a portmanteau of BEst subSET, which is the common thread
underlying the functions in this package: to explore large multivariate data
sets and extract useful features from spaces where many, if not most, of the
features are expected to be redundant or irrelevant. \code{beset} brings
together predictive modeling methods from other packages that perform
variable selection either explicitly (e.g., best subset selection) or
implicitly (e.g., lasso regression) as a part of model fitting.
}

