% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beset_elnet.R
\name{beset_elnet}
\alias{beset_elnet}
\title{Beset GLM with Elasticnet Regularization}
\usage{
beset_elnet(form, train_data, standardize = TRUE, method = c("best", "1SE"),
  max_features = 1000, alpha = c(0.05, 0.5, 0.95), n_folds = 5,
  n_repeats = 5, seed = 42, n_cores = NULL)
}
\arguments{
\item{form}{A formula of the form y ~ x1 + x2 + ...}

\item{train_data}{Data frame containing the training data set and all of the
variables specified in \code{form}.}

\item{alpha}{Numeric vector of alpha values between 0 and 1 to use as tuning
parameters. \code{alpha = 0} results in ridge regression, and \code{alpha =
1} results in lasso regression. Values in between result in a mixture of L1
and L2 penalties. (Values closer to 0 weight the L2 penalty more heavily,
and values closer to 1 weight the L1 penalty more heavily.)}

\item{n_folds}{Integer indicating the number of cross-validation folds.}

\item{n_repeats}{Number of times cross-validation should be repeated.}

\item{seed}{Seed for random number generator.}

\item{n_cores}{Number of cores to use for parallel execution. If not
specified, the number of cores is set to the value of
\code{options("cores")}, if specified, or to approximately half the number of
cores detected by the \code{parallel} package. To determine the theoretical
maximum number of cores you have available, see
\code{\link[parallel]{detectCores}}, but note that the actual number of cores
available may be less. See \href{
https://stat.ethz.ch/R-manual/R-devel/library/parallel/doc/parallel.pdf}{
\code{parallel} package vignette} for more information.}
}
\description{
\code{beset_elnet} is a wrapper to \code{\link[glmnet]{glmnet}} for fitting
 generalized linear modesl via penalized maximum likelihood, providing
 automated data preprocessing and selection of both the elasticnet penalty
 and regularization parameter through cross-validation.
}
\seealso{
\code{\link[glmnet]{glmnet}}, \code{\link[caret]{train}}
}

