% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beset_glm.R
\name{beset_glm}
\alias{beset_glm}
\title{Best Subset Selection for Generalized Linear Models}
\usage{
beset_glm(form, train_data, test_data = NULL, family = "gaussian",
  link = NULL, ..., p_max = 10, n_folds = 10, n_repeats = 10,
  oneSE = TRUE, n_cores = NULL, seed = 42)
}
\arguments{
\item{form}{A model \code{\link[stats]{formula}}.}

\item{train_data}{A \code{\link[base]{data.frame}} with the variables in
\code{form} and the data to be used in model training.}

\item{test_data}{Optional \code{\link[base]{data.frame}} with the variables
in \code{form} and the data to be used in model testing.}

\item{family}{Character string naming the error distribution to be used in
the model. Available families are listed under 'List of available families
and link functions'.}

\item{link}{Optional character string naming the link function to be used in
the model. Available links and their defaults differ by \code{family} and are
listed under 'List of available families and link functions'.}

\item{...}{Additional arguments to be passed to \code{\link[stats]{glm}}}

\item{p_max}{Maximum number of predictors to attempt to fit.}

\item{n_folds}{Integer indicating the number of folds to use for
cross-validation.}

\item{n_repeats}{Integer indicating the number of times cross-validation
should be repeated.}

\item{n_cores}{Optional integer value indicating the number of workers to run
in parallel during subset search and cross-validation. By default, this will
be set to equal half the detectable cores on your machine. You may wish to
change this depending on your hardware and OS.
See \code{\link[parallel]{parallel-package}} for more information.}

\item{seed}{An integer used to seed the random number generator when
assigning observations to folds.}
}
\value{
A "beset_glm" object with the following components:
\enumerate{
 \item\describe{
   \item{best_model}{an object of class \code{\link[stats]{glm}}
   corresponding to the best model}
   }
 \item\describe{
   \item{all_subsets}{a data frame containing fit statistics for every
     possible combination of predictors:
     \describe{
     \item{n_pred}{the number of predictors in model}
     \item{form}{formula for model}
     \item{train_CE}{Cross entropy between model predictions and
       \code{train_data}}
     \item{test_CE}{if \code{test_data} is provided, the cross entropy
       between the model fit to \code{train_data} and the \code{test_data}}
      }
   }
 }
 \item\describe{
   \item{best_subsets}{a data frame containing cross-validation statistics
     for the best model for each \code{n_pred} listed in \code{all_subsets}.
     In addition to the columns found in \code{all_subsets}, contains the
     following:
     \describe{
     \item{cv_CE}{the mean cross entropy between the predictions of models
       fit to \code{n-1} folds and the left-out fold}
     \item{cv_CE_SE}{the standard error of the mean cross entropy}
      }
   }
 }
}
}
\description{
\code{beset_glm} performs best subset selection using repeated
cross-validation to find the optimal number of predictors for several
families of generalized linear models.
}
\details{
\code{beset_glm} performs best subset selection for generalized linear
models, fitting a separate model for each possible combination of predictors,
i.e., all models that contain exactly 1 predictor, all models that contain
exactly 2 predictors, and so forth. For each number of predictors,
\code{beset_glm} picks the model with the maximum likelihood and then
estimates how well this model predicts new data using \code{k}-fold
cross-validation, i.e., how well a model trained using \eqn{k - 1} folds
predicts the left-out fold.
}
\section{Cross-validation details}{

\code{beset_glm} uses \code{\link[caret]{createFolds}} to randomly assign
observations to \code{k} folds within levels of the outcome when the outcome
is a factor or within subgroups based on percentiles when the outcome is
numeric. This insures that the folds will be matched in terms of the
outcome's frequency distribution. \code{beset_glm} also insures the
reproducibility of your analysis by requiring a \code{seed} to the random
number generator as one of its arguments.
}

\section{List of available families and link functions}{

\describe{
 \item{\code{"gaussian"}}{The Gaussian family accepts the links
   \code{"identity"} (default), \code{"log"}, and \code{"inverse"}.}
 \item{\code{"binomial"}}{The binomial family accepts the links
   \code{"logit"} (default), \code{"probit"}, \code{"cauchit"}, \code{"log"}, and
   \code{"cloglog"} (complementary log-log).}
 \item{\code{"poisson"}}{The Poisson family accepts the links \code{"log"}
   (default), \code{"sqrt"}, and \code{"identity"}.}
 \item{\code{"negbin"}}{The negative binomial family accepts the links
   \code{"log"} (default), \code{"sqrt"}, and \code{"identity"}.}
 }
}

\section{Warnings}{

\enumerate{
 \item \code{beset_glm} handles missing data by performing listwise deletion.
  No other options for handling missing data are provided at this time. The
  user is encouraged to deal with missing values prior to running this
  function.
 \item \code{beset_glm} is intended for use with additive models only.
   While there is no prohibition against the inclusion of interaction or
   polynomial terms, this practice is strongly discouraged. At best, this
   will result in an inefficient search because \code{beset_glm} performs an
   exhaustive search over all possible variable subsets, including subsets
   that are hierarchically incomplete, i.e., subsets that contain an
   interaction term but are missing one or more of the subterms that comprise
   it. At worst, it may return one of these hierarchically incomplete models
   as the best model, an undesirable result if one cares about
   interpretability. If one wishes the model search to include interaction
   and/or non-linear effects, the
   \href{https://cran.r-project.org/web/packages/earth/index.html}{MARS}
   technique is recommended instead. \item \code{beset_glm} is best suited
   for searching over a small number of predictors (less than 10). For a
   large number of predictors (more than 20), \code{\link{beset_elnet}} is
   recommended instead. However, note that \code{\link{beset_elnet}} only
   works with a more restricted set of distributions.
}
}
\seealso{
\code{\link[caret]{createFolds}}, \code{\link[stats]{glm}},
\code{\link[base]{set.seed}}, \code{\link[MASS]{glm.nb}}
}

