% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beset_glm.R
\name{beset_glm}
\alias{beset_glm}
\alias{beset_glm}
\alias{beset_lm}
\title{Best Subset Selection for Generalized Linear Models}
\usage{
beset_glm(form, data, test_data = NULL, p_max = 10, family = "gaussian",
  link = NULL, ..., n_cores = 2, n_folds = 10, n_repeats = 10,
  seed = 42)

beset_lm(form, data, test_data = NULL, p_max = 10, n_cores = 2,
  n_folds = 10, n_repeats = 10, seed = 42)
}
\arguments{
\item{form}{A model \code{\link[stats]{formula}}.}

\item{data}{Data frame with the variables in \code{form} and the data
to be used for model fitting.}

\item{test_data}{Optional data frame with the variables in \code{form} and
the data to be used for model validation.}

\item{p_max}{Maximum number of predictors to attempt to fit.}

\item{family}{Character string naming the error distribution to be used in
the model. Available families are listed under 'List of available families
and link functions'.}

\item{link}{Optional character string naming the link function to be used in
the model. Available links and their defaults differ by \code{family} and are
listed under 'List of available families and link functions'.}

\item{...}{Additional arguments to be passed to \code{\link[stats]{glm}}}

\item{n_cores}{Integer value indicating the number of workers to run in
parallel during subset search and cross-validation. By default, this will
be set to 2. To determine the theoretical maximum number of cores you have
available, see \code{\link[parallel]{detectCores}}, but note that the actual
number of cores available may be less. See
\code{\link[parallel]{parallel-package}} for more information.}

\item{n_folds}{Integer indicating the number of folds to use for
cross-validation.}

\item{n_repeats}{Integer indicating the number of times cross-validation
should be repeated.}

\item{seed}{An integer used to seed the random number generator when
assigning observations to folds.}
}
\value{
A "beset_glm" object with the following components:
\enumerate{
 \item\describe{
   \item{best_aic}{an object of class \code{\link[stats]{glm}}
   corresponding to the model with the lowest Akaike Information Criterion}
 }
 \item\describe{
   \item{cv_params}{list of the values of the parameters used for
   cross-validation, i.e., \code{n_folds}, \code{n_repeats}, \code{seed}}
 }
  \item\describe{
    \item{model_data}{data frame used to identify best subsets}
 }
 \item\describe{
   \item{stats}{a list with three data frames:
     \describe{
       \item{fit}{statistics for every possible combination of predictors:
         \describe{
           \item{n_pred}{the total number of predictors in model; note that
              the number of predictors for a factor variable corresponds to the
              number of factor levels minus 1}
           \item{form}{formula for model}
           \item{aic}{\eqn{-2*log-likelihood + k*npar}, where \eqn{npar}
             represents the number of parameters in the fitted model, and
             \eqn{k = 2}}
           \item{dev}{twice the difference between the log-likelihoods of the
             saturated and fitted models, multiplied by the scale parameter}
           \item{mae}{mean absolute error}
           \item{mce}{mean cross entropy, estimated as
             \eqn{-log-likelihood/N}, where \eqn{N} is the number of
             observations}
           \item{mse}{mean squared error}
           \item{r2}{R-squared, calculated as
             \eqn{1 - deviance/null deviance}}
           }
         }
     \item{cv}{a data frame containing cross-validation statistics
     for the best model for each \code{n_pred} listed in \code{fit_stats}.
     Each metric is computed using \code{\link{predict_metrics}}, with
     models fit to \eqn{n-1} folds and predictions made on the left-out fold.
     Each metric is followed by its standard error. The data frame
     is otherwise the same as that documented for \code{fit}, except
     AIC is omitted.}
     \item{test_stats}{if \code{test_data} is provided, a data frame
     containing prediction metrics for the best model for each \code{n_pred}
     listed in \code{fit} as applied to the \code{test_data}.}
      }
    }
  }
}
}
\description{
\code{beset_glm} performs best subset selection using repeated
cross-validation to find the optimal number of predictors for several
families of generalized linear models.
}
\details{
\code{beset_glm} performs best subset selection for generalized linear
models, fitting a separate model for each possible combination of predictors
(all models that contain exactly 1 predictor, all models that contain
exactly 2 predictors, and so forth). For each number of predictors,
\code{beset_glm} picks the model with the maximum likelihood and then
estimates how well this model predicts new data using \code{k}-fold
cross-validation (how well a model trained using \eqn{k - 1} folds
predicts the left-out fold).
}
\section{Cross-validation details}{

\code{beset_glm} uses \code{\link[caret]{createMultiFolds}} to randomly
partition the data set into \code{n_folds} * \code{n_repeats} folds within
strata (factor levels for factor outcomes, percentile-based groups for
numeric outcomes). This insures that the folds will be matched in terms of
the outcome's frequency distribution. \code{beset_glm} also insures the
reproducibility of your analysis by requiring a \code{seed} to the random
number generator as one of its arguments.
}

\section{List of available families and link functions}{

\describe{
 \item{\code{"gaussian"}}{The Gaussian family accepts the links
   \code{"identity"} (default), \code{"log"}, and \code{"inverse"}.}
 \item{\code{"binomial"}}{The binomial family accepts the links
   \code{"logit"} (default), \code{"probit"}, \code{"cauchit"}, \code{"log"}, and
   \code{"cloglog"} (complementary log-log).}
 \item{\code{"poisson"}}{The Poisson family accepts the links \code{"log"}
   (default), \code{"sqrt"}, and \code{"identity"}.}
 \item{\code{"negbin"}}{The negative binomial family accepts the links
   \code{"log"} (default), \code{"sqrt"}, and \code{"identity"}.}
 }
}

\section{Warnings}{

\enumerate{
 \item \code{beset_glm} handles missing data by performing listwise deletion.
  No other options for handling missing data are provided at this time. The
  user is encouraged to deal with missing values prior to running this
  function.
 \item \code{beset_glm} is intended for use with additive models only.
   While there is no prohibition against the inclusion of interaction or
   polynomial terms, this practice is strongly discouraged. At best, this
   will result in an inefficient search because \code{beset_glm} performs an
   exhaustive search over all possible variable subsets, including subsets
   that are hierarchically incomplete, i.e., subsets that contain an
   interaction term but are missing one or more of the subterms that comprise
   it. At worst, it may return one of these hierarchically incomplete models
   as the best model, an undesirable result if one cares about
   interpretability. If one wishes the model search to include interaction
   and/or non-linear effects, the
   \href{https://cran.r-project.org/web/packages/earth/index.html}{MARS}
   technique is recommended instead. \item \code{beset_glm} is best suited
   for searching over a small number of predictors (less than 10). For a
   large number of predictors (more than 20), \code{\link{beset_elnet}} is
   recommended instead. However, note that \code{\link{beset_elnet}} only
   works with a more restricted set of distributions.
}
}

\seealso{
\code{\link[caret]{createFolds}}, \code{\link[stats]{glm}},
\code{\link[base]{set.seed}}, \code{\link[MASS]{glm.nb}}
}
