% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beset_zeroinfl.R
\name{beset_zeroinfl}
\alias{beset_zeroinfl}
\title{Best Subset Selection for Zero-inflated Count Data Regression}
\usage{
beset_zeroinfl(form, train_data, test_data = NULL, family = "poisson",
  link = "logit", ..., p_count_max = 10, p_zero_max = 10, n_folds = 10,
  n_repeats = 10, n_cores = NULL, seed = 42)
}
\arguments{
\item{form}{A model \code{\link[stats]{formula}}.}

\item{train_data}{A \code{\link[base]{data.frame}} with the variables in
\code{form} and the data to be used in model training.}

\item{test_data}{Optional \code{\link[base]{data.frame}} with the variables
in \code{form} and the data to be used in model testing.}

\item{family}{Character string naming the count model family. Options are
\code{"poisson"} (default), \code{"negbin"}, or \code{"geometric"}. (Note a
log link is always used with the count model).}

\item{link}{Character string naming the link function in the binary
zero-inflation model. Options are \code{"logit"} (default), \code{"probit"},
\code{"cloglog"}, \code{"cauchit"}, or \code{"log"}. (Note a binomial family
is always used with the zero-inflation model).}

\item{...}{Optional named arguments accepted by \code{\link[pscl]{zeroinfl}}.}

\item{n_folds}{Integer indicating the number of folds to use for
cross-validation.}

\item{n_repeats}{Integer indicating the number of times cross-validation
should be repeated.}

\item{n_cores}{Optional integer value indicating the number of workers to run
in parallel during subset search and cross-validation. By default, this will
be set to equal half of the detectable cores on your machine. You may wish to
change this depending on your hardware and OS.
See \code{\link[parallel]{parallel-package}} for more information.}

\item{seed}{An integer used to seed the random number generator when
assigning observations to folds.}
}
\value{
A "beset_zeroinfl" object with the following components:
\enumerate{
 \item\describe{
   \item{best_model}{an object of class \code{\link[pscl]{zeroinfl}}
   corresponding to the best model with the number of parameters with the
   smallest cross-validation error (largest cross-validation R-squared)}
   }
 \item\describe{
   \item{best_model_1SE}{an object of class \code{\link[pscl]{zeroinfl}}
   corresponding to the best model with the smallest number of
   parameters within one standard error of the smallest cross-validation
   error (largest cross-validation R-squared)}
   }
 \item\describe{
   \item{all_count_subsets}{a data frame containing fit statistics for every
     possible combination of predictors:
     \describe{
     \item{n_count_pred}{the number of predictors in model}
     \item{form}{formula for model}
     \item{train_CE}{Proportion of variance or deviance in the
       \code{train_data} explained by each size of best model}
     \item{test_CE}{if \code{test_data} is provided, the R-squared when
       the model fit to \code{train_data} is applied to the \code{test_data}}
      }
   }
 }
  \item\describe{
   \item{all_zero_subsets}{a data frame containing fit statistics for every
     possible combination of predictors:
     \describe{
     \item{n_zero_pred}{the number of predictors in model}
     \item{form}{formula for model}
     \item{train_CE}{Proportion of variance or deviance in the
       \code{train_data} explained by each size of best model}
     \item{test_CE}{if \code{test_data} is provided, the R-squared when
       the model fit to \code{train_data} is applied to the \code{test_data}}
      }
   }
 }
 \item\describe{
   \item{best_subsets}{a data frame containing cross-validation statistics
     for the best model for each combination of \code{n_count_pred} and
       \code{n_zero_pred}:
     \describe{
     \item{n_count_pred}{the number of predictors in model}
     \item{n_zero_pred}{the number of predictors in model}
     \item{form}{formula for best model of \code{n_count_pred} and
       \code{n_zero_pred}}
     \item{train_CE}{Proportion of variance or deviance in the
       \code{train_data} explained by each size of best model}
     \item{test_CE}{if \code{test_data} is provided, the R-squared when
       the model fit to \code{train_data} is applied to the \code{test_data}}
     \item{cv_CE}{the mean cross-validation R-squared for each size of best
       model, i.e., on average, how well models fit to \code{n-1} folds
       explain the left-out fold}
     \item{cv_CE_SE}{the standard error of the cross-validation R-squared for
      each size of best model}
      }
     }
   }
 }
}
\description{
\code{beset_zeroinfl} performs best subset selection using repeated
cross-validation to find the optimal number of predictors for zero-inflated
regression models.
}
\details{
\code{beset_zeroinfl} performs best subset selection for zero-inflated
count models, fitting both a count model (e.g., negative binomial regression)
and zero-inflation model (e.g., logistic regression) for every possible
combination of predictors. For models with the same number of parameters,
e.g., \code{m} predictors of the count process + \code{n} predictors of the
zero-inflation process, \code{beset_zeroinfl} picks the model with the
maximum log-likelihood. This results in a best fit for every \code{m * n}
possible number of predictors. \code{beset_zeroinfl} then uses \code{k}-fold
cross-validation to select the "best of the best": the best model with the
\code{m * n} number of predictors that minimizes prediction error, i.e., how
well the best models trained using \eqn{k - 1} folds predict the out-of-fold
sample.

\code{beset_zeroinfl} uses \code{\link[caret]{createFolds}} to randomly
assign observations to \code{k} folds within subgroups based on percentiles
of the outcome is numeric. This insures that the folds will be matched in
terms of the outcome's frequency distribution. \code{beset_zeroinfl} also
insures the reproducibility of your analysis by requiring a \code{seed} to
the random number generator as one of its arguments.
}
\section{Warnings}{

\enumerate{
 \item \code{beset_zeroinfl} handles missing data by performing listwise
   deletion. No other options for handling missing data are provided at this
   time. The user is encouraged to deal with missing values prior to running
   this function.
 \item \code{beset_zeroinfl} is intended for use with additive models only.
   While there is no prohibition against the inclusion of interaction or
   polynomial terms, this practice is strongly discouraged. At best, this
   will result in an inefficient search because \code{beset_zeroinfl}
   performs an exhaustive search over all possible variable subsets,
   including subsets that are hierarchically incomplete, i.e. subsets that
   contain an interaction term but are missing one or more of the subterms
   that comprise it. At worst, it may return one of these hierarchically
   incomplete models as the best model, an undesirable result if one cares
   about interpretability. If one wishes the model search to include
   interaction and/or non-linear effects, the
   \href{https://cran.r-project.org/web/packages/earth/index.html}{MARS}
   technique is recommended instead. \item \code{beset_zeroinfl} is best
   suited for searching over a small number of predictors (less than 10). For
   a large number of predictors (more than 20), \code{\link{beset_elnet}} is
   recommended instead. However, note that \code{\link{beset_elnet}} only
   works with a more restricted set of distributions.
}
}
\seealso{
\code{\link[caret]{createFolds}}, \code{\link{beset_lm}},
\code{\link[pscl]{zeroinfl}}, \code{\link[base]{set.seed}},
\code{\link{cross_entropy}}
}

