% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beset_zeroinfl.R
\name{beset_zeroinfl}
\alias{beset_zeroinfl}
\title{Best Subset Selection for Generalized Linear Models}
\usage{
beset_zeroinfl(form, train_data, test_data = NULL, family = "poisson",
  link = "logit", ..., search = "stagewise", p_max = 10, n_folds = 10,
  n_repeats = 10, n_cores = NULL, seed = 42)
}
\arguments{
\item{family}{Character string naming the count model family. Options are
\code{"poisson"} (default), \code{"negbin"}, or \code{"geometric"}. (Note a
log link is always used with the count model).}

\item{link}{Character string naming the link function in the binary
zero-inflation model. Options are \code{"logit"} (default), \code{"probit"},
\code{"cloglog"}, \code{"cauchit"}, or \code{"log"}. (Note a binomial family
is always used with the zero-inflation model).}

\item{...}{Optional named arguments accepted by \code{\link[pscl]{zeroinfl}}.}

\item{search}{Character string naming the search strategy to use. Options are
\code{"stagewise"} (default), \code{"identical"}, or \code{"exhasutive"}. See
'Details' for explanation.}
}
\value{
A list with the following three components:
\enumerate{
 \item\describe{
   \item{best_model}{an object of class \code{"zeroinfl"} corresponding
   to the best model with the number of parameters with the smallest
   cross-validation error (largest cross-validation R-squared)}
   }
 \item\describe{
   \item{best_model_1SE}{an object of class \code{"zeroinfl"}
   corresponding to the best model with the smallest number of
   parameters within one standard error of the smallest cross-validation
   error (largest cross-validation R-squared)}
   }
 \item\describe{} \code{\link[base]{data.frame}} with the following columns:
   \describe{
   \item{k_folds}{the number of k-folds used}
   \item{n_preds}{the number of predictors in model}
   \item{R2_train}{the mean in-fold (training) R-squared for each size of
     best model}
   \item{R2_train_SE}{the standard error of the mean in-fold (training)
     R-squared for each size of best model}
   \item{R2_cv}{the mean out-of-fold (test) R-squared for each size of best
     model}
   \item{R2_cv_SE}{the standard error of the mean out-of-fold (test)
     R-squared for each size of best model}
   \item{R2_test}{if \code{test_data} is provided, the mean R-squared when
     the model fit to each subsample of the training data is applied to the
     independent test set}
   \item{R2_test_SE}{if \code{test_data} is provided, the standard error of
     the mean R-squared when the model fit to each subsample of the training
     data is applied to the independent test set}
   }
 }
}
\description{
\code{beset_zeroinfl} performs best subset selection using repeated
cross-validation to find the optimal number of predictors for zero-inflated
regression models.
}
\details{
\code{beset_zeroinfl} performs best subset selection for zero-inflated
count models, fitting both a count model (e.g., negative binomial regression)
and zero-inflation model (e.g., logistic regression) for every possible
combination of predictors. For models with the same number of parameters,
e.g., \code{m} predictors of the count process + \code{n} predictors of the
zero-inflation process, \code{beset_zeroinfl} picks the model with the
maximum log-likelihood. This results in a best fit for every \code{m * n}
possible number of predictors. \code{beset_zeroinfl} then uses \code{k}-fold
cross-validation to select the "best of the best": the best model with the
\code{m * n} number of predictors that minimizes prediction error, i.e., how
well the best models trained using \eqn{k - 1} folds predict the out-of-fold
sample.

\code{beset_zeroinfl} uses \code{\link[caret]{createFolds}} to randomly
assign observations to \code{k} folds within subgroups based on percentiles
of the outcome is numeric. This insures that the folds will be matched in
terms of the outcome's frequency distribution. \code{beset_zeroinfl} also
insures the reproducibility of your analysis by requiring a \code{seed} to
the random number generator as one of its arguments.
}
\section{Warnings}{

\enumerate{
 \item \code{beset_zeroinfl} handles missing data by performing listwise
   deletion. No other options for handling missing data are provided at this
   time. The user is encouraged to deal with missing values prior to running
   this function.
 \item \code{beset_zeroinfl} is intended for use with additive models only.
   While there is no prohibition against the inclusion of interaction or
   polynomial terms, this practice is strongly discouraged. At best, this
   will result in an inefficient search because \code{beset_zeroinfl}
   performs an exhaustive search over all possible variable subsets,
   including subsets that are hierarchically incomplete, i.e. subsets that
   contain an interaction term but are missing one or more of the subterms
   that comprise it. At worst, it may return one of these hierarchically
   incomplete models as the best model, an undesirable result if one cares
   about interpretability. If one wishes the model search to include
   interaction and/or non-linear effects, the
   \href{https://cran.r-project.org/web/packages/earth/index.html}{MARS}
   technique is recommended instead. \item \code{beset_zeroinfl} is best
   suited for searching over a small number of predictors (less than 10). For
   a large number of predictors (more than 20), \code{\link{beset_elnet}} is
   recommended instead. However, note that \code{\link{beset_elnet}} only
   works with a more restricted set of distributions.
}
}
\seealso{
\code{\link[caret]{createFolds}}, \code{\link{beset_lm}},
\code{\link[pscl]{zeroinfl}}, \code{\link[base]{set.seed}},
\code{\link{predict_R2}}
}

