% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\alias{compare.cross_valid}
\alias{compare.data.frame}
\alias{compare.numeric}
\alias{compare.beset}
\title{Compare Predictive Performance of Two Models}
\usage{
compare(yhat1, yhat2, ...)

\method{compare}{cross_valid}(yhat1, yhat2, n_rep = 1000, conf = 0.95,
  parallel_type = NULL, n_cores = NULL, cl = NULL)

\method{compare}{data.frame}(yhat1, yhat2, y, n_rep = 1000, conf = 0.95,
  family = "gaussian", phi = NULL, theta = NULL, parallel_type = NULL,
  n_cores = NULL, cl = NULL, ...)

\method{compare}{numeric}(yhat1, yhat2, y, ...)

\method{compare}{beset}(yhat1, yhat2, ...)
}
\arguments{
\item{yhat1}{A data frame consisting of cross-validated predictions from a
benchmark model, an object containing such a data frame, e.g., a
"cross_valid" object returned by \code{\link{validate}}, or an object that
can be passed to \code{\link{validate}}.}

\item{yhat2}{An object of the same type as \code{yhat1} to be compared}

\item{n_rep}{\code{Integer} giving the number of bootstrap replicates to
perform for each repetition of cross-validated predictions. For example,
if \code{yhat1} and \code{yhat2} contain 10 columns of predictions, the
default value of \code{n_rep = 1000} will result in
\eqn{1000 \times 10 = 10,000} replicates total.}

\item{conf}{Confidence level for the difference between model performance.}

\item{parallel_type}{(Optional) character string indicating the type of
parallel operation to be used, either \code{"fork"} or \code{"sock"}. If
omitted and \code{n_cores > 1}, the default is \code{"sock"} for Windows and
\code{"fork"} for any other OS.}

\item{n_cores}{Integer value indicating the number of workers to run in
parallel during subset search and cross-validation. By default, this will
be set to one fewer than the maximum number of physical cores you have
available, as indicated by \code{\link[parallel]{detectCores}}. Set to 1 to
disable parallel processing.}

\item{cl}{(Optional) \code{\link[parallel]{parallel}} or
\code{\link[snow]{snow}} cluster for use if \code{parallel_type = "sock"}.
If not supplied, a cluster on the local machine is automatically created.}

\item{family}{Character string naming the error distribution to be used in
the model. Available families are listed under 'List of available families
and link functions'.}
}
\description{
Compare Predictive Performance of Two Models
}
\section{Methods (by class)}{
\itemize{
\item \code{cross_valid}: S3 method for class 'cross_valid'

\item \code{data.frame}: S3 method for class 'data.frame'

\item \code{numeric}: S3 method for class 'numeric'

\item \code{beset}: two 'beset' class models
}}

