% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_list.R
\name{cor_list}
\alias{cor_list}
\title{Construct a Correlation List}
\usage{
cor_list(cor_matrix)
}
\arguments{
\item{cor_matrix}{Correlation matrix}
}
\value{
A \code{cor_list} object with the following three vectors:
 \describe{
   \item{\code{x1}}{variables on one side of the correlation}
   \item{\code{x2}}{variables on the other side of the correlation}
   \item{\code{coef}}{the correlation coefficient between \code{x1} and
     \code{x2}}
 }
}
\description{
Construct a \code{cor_list} object from a correlation matrix.
}
\details{
The R stats package returns bivariate correlations as a matrix of values,
which can be difficult to parse when there are many correlations.
\code{cor_list} converts the correlation matrix into a list of all
row-column pairs, removing the diagonal entries.
}

