% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{dependence}
\alias{dependence}
\alias{dependence.nested}
\alias{dependence.beset}
\alias{dependence.beset_rf}
\alias{plot.part_depend}
\title{Partial-Dependence Effects}
\usage{
dependence(object, ...)

\method{dependence}{nested}(object, x = NULL, cond = NULL, alpha = NULL,
  lambda = NULL, n_pred = NULL, metric = "auto", oneSE = TRUE,
  x_lab = NULL, y_lab = NULL, ..., parallel_type = NULL, n_cores = NULL,
  cl = NULL)

\method{dependence}{beset}(object, x = NULL, cond = NULL, alpha = NULL,
  lambda = NULL, n_pred = NULL, metric = "auto", oneSE = TRUE,
  x_lab = NULL, y_lab = NULL, ...)

\method{dependence}{beset_rf}(object, x = NULL, cond = NULL, x_lab = NULL,
  y_lab = NULL, ..., parallel_type = NULL, n_cores = NULL, cl = NULL)

\method{plot}{part_depend}(x, order = "delta", p_max = 16, ...)
}
\arguments{
\item{object}{A "beset" model object for which partial dependence plots
are desired.}

\item{...}{Additional arguments affecting the plots produced.}

\item{x}{(Optional) \code{character} vector giving name(s) of predictors
for which partial dependence effects should be estimated. If omitted,
partial dependence effects will be estimated for all predictors in the
model.}

\item{cond}{(Optional) named \code{list} of the values to use for one or more
predictors when estimating the partial effect of the predictor(s) named in
\code{x}. Any predictor omitted from this list will be assigned the mean
observed value for continuous variables, or the modal level for factors.}

\item{n_pred}{(Optional) \code{integer} number of predictors that the best
model should contain. If specified, all other arguments are ignored.}

\item{metric}{\code{Character} string giving prediction metric on which to
base model selection. Can be one of \code{"auc"} for area under the (ROC)
curve (only available for binomial family), \code{"mae"} for mean absolute
error (not available for binomial family), \code{"mae"} for mean absolute
error, \code{"mce"} for mean cross entropy, or \code{"mse"} for mean squared
error. Default is \code{"auto"} which plots MSE for Gaussian-family models
and MCE for all other families.}

\item{oneSE}{\code{Logical} indicating whether or not to use the "one
standard error" rule. If \code{TRUE} (default) the simplest model within one
standard error of the optimal model is returned. If \code{FALSE} the model
with the optimal cross-validation performance is returned.}

\item{x_lab}{(Optional) \code{character} vector giving replacement labels
for \code{x}. Labels should be listed in the same order as the predictor
names in \code{x}, or, if \code{x} is omitted, the same column order as the
predictors in the model data frame. If omitted, plots will be labeled with
the names of the predictors as they appear in the model object.}

\item{y_lab}{(Optional) \code{character} string giving replacement label
for the response variable in the model. If omitted, plots will be labeled
with the name of the response as it appears in the model object.}

\item{parallel_type}{(Optional) character string indicating the type of
parallel operation to be used, either \code{"fork"} or \code{"sock"}. If
omitted and \code{n_cores > 1}, the default is \code{"sock"} for Windows and
otherwise either \code{"fork"} or \code{"sock"} depending on which process is
being run.}

\item{n_cores}{Integer value indicating the number of workers to run in
parallel during subset search and cross-validation. By default, this will
be set to one fewer than the maximum number of physical cores you have
available, as indicated by \code{\link[parallel]{detectCores}}. Set to 1 to
disable parallel processing.}

\item{cl}{(Optional) \code{\link[parallel]{parallel}} or
\code{\link[snow]{snow}} cluster for use if \code{parallel_type = "sock"}.
If not supplied, a cluster on the local machine is automatically created.}

\item{order}{If plotting a grid, order in which partial dependence plots
should appear. Options are \code{"delta"} (default), which arranges plots
based on the magnitude of change in \code{y} over the range of each \code{x},
and \code{"import"}, which arranges plots based on the variable importance
scores returned by the object's \code{\link{importance}} method.}

\item{p_max}{Maximum number of partial dependence effects to plot.
Default is 16, resulting in a 4 X 4 grid.}
}
\description{
Generate data for partial dependence plots.
}
