% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_nb.R
\name{glm_nb}
\alias{glm_nb}
\title{Fit a Negative Binomial Generalized Linear Model}
\usage{
glm_nb(formula, data, weights, subset, na.action, start = NULL, etastart,
  mustart, control = stats::glm.control(...), method = "glm.fit",
  model = FALSE, x = FALSE, y = FALSE, contrasts = NULL, ...,
  init.theta, link = "log")
}
\description{
A modification of \code{\link[MASS]{glm.nb}} that can be programmed with.
The main change is to avoid non-standard evaluation of the link argument so
that this argument is passed as a "quoted" string, which requires the
removal of the line `link <- substitute(link)`. This in turn enables the
removal of `do.call` syntax. Other edits were done to define the namespaces
of other MASS functions that are called by `glm.nb`, to change the default
arguments to avoid returning copies of the model data and the response
vector, and to return $data to maintain consistency with the object returned
by \code{\link[stats]{glm}}.
}
