% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.beset}
\alias{plot.beset}
\alias{plot.beset_elnet}
\alias{plot.beset_glm}
\alias{plot.beset_zeroinfl}
\title{Plot Methods for \code{beset} Objects}
\usage{
\method{plot}{beset_elnet}(object, type = "cv", metric = "MCE")

\method{plot}{beset_glm}(object, metric = "MCE")

\method{plot}{beset_zeroinfl}(object, type = "cv", metric = "MCE")
}
\arguments{
\item{object}{An object of class \code{"beset_glm"},
\code{"beset_zeroinfl"}, or \code{"beset_elnet"}}

\item{type}{Type of error to plot. Can be one of \code{"cv"}, \code{"train"},
or \code{"test"}.}

\item{metric}{Which error metric to plot. Can be one of \code{"MCE"} for
mean cross entropy, \code{"MSE"} for mean squared error, or \code{"R2"} for
R-squared.}
}
\description{
Plot Methods for \code{beset} Objects
}
