% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_plot.R
\name{plot.beset}
\alias{plot.beset}
\title{Plot Methods for \code{beset} Objects}
\usage{
\method{plot}{beset}(x, metric = "auto", se = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{"beset_glm"},
\code{"beset_zeroinfl"}, or \code{"beset_elnet"}}

\item{metric}{\code{Character} string giving the performance metric to plot.
Can be one of \code{"auc"} for area under the (ROC) curve (only available for
binomial family), \code{"mae"} for mean absolute error (not available for
binomial family), \code{"mce"} for mean cross entropy, \code{"mse"} for
mean squared error, or \code{"rsq"} for R-squared. Default is \code{"auto"}
which plots MSE for Gaussian-family models and MCE for all other families.}

\item{se}{\code{Logical} indicating whether or not to plot error bars.
\code{TRUE} by default.}
}
\description{
Plot Methods for \code{beset} Objects
}
