% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.beset_glm.R
\name{plot.beset_glm}
\alias{plot.beset_glm}
\title{Plot Method for the \code{beset_glm} Class}
\usage{
\method{plot}{beset_glm}(object, metric = "MCE", se = TRUE, title = "")
}
\arguments{
\item{object}{An object of class \code{beset_glm}.}

\item{title}{Optional character vector giving plot title.}

\item{SE}{Logical indicating whether or not standard-error bars should be
plotted.}
}
\description{
\code{plot.beset_glm} takes the output of a \code{\link{beset_glm}} object
and creates a line plot using \code{\link[ggplot2]{ggplot}}.
}
\details{
\code{plot.beset_glm} produces a plot showing cross-validated prediction
error as a function of the number of model parameters from a
\code{\link{beset_glm}} object. Line graphs are plotted for the training,
cross-validated, and, if included, independent-test error, which are labeled
in the legend as "Train", "CV", and "Test", respectively.
}
