% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_summary.R
\name{summary.beset}
\alias{summary.beset}
\title{Summary Methods for \code{beset} Objects}
\usage{
\method{summary}{beset}(object, n_pred = NULL, alpha = NULL,
  lambda = NULL, metric = "auto", oneSE = TRUE, ...)
}
\arguments{
\item{object}{An object for which a summary is desired.}

\item{n_pred}{(Optional) \code{integer} number of predictors that the best
model should contain. If specified, all other arguments are ignored.}

\item{metric}{\code{Character} string giving prediction metric on which to
base model selection. Can be one of \code{"auc"} for area under the (ROC)
curve (only available for binomial family), \code{"mae"} for mean absolute
error (not available for binomial family), \code{"mae"} for mean absolute
error, \code{"mce"} for mean cross entropy, or \code{"mse"} for mean squared
error. Default is \code{"auto"} which plots MSE for Gaussian-family models
and MCE for all other families.}

\item{oneSE}{\code{Logical} indicating whether or not to use the "one
standard error" rule. If \code{TRUE} (default) the simplest model within one
standard error of the optimal model is returned. If \code{FALSE} the model
with the optimal cross-validation performance is returned.}

\item{...}{Additional arguments passed to model summary methods.}

\item{robust}{\code{Logical} indicating whether or not to produce summary
output for robust predictors, defined here as having non-zero beta
coefficients for every repetition of the cross-validation procedure}
}
\description{
These functions are all methods for class \code{beset} objects.
}
