% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary.beset}
\alias{summary.beset}
\alias{summary.beset_elnet}
\alias{summary.beset_glm}
\alias{summary.beset_zeroinfl}
\title{Summary Methods for \code{beset} Objects}
\usage{
\method{summary}{beset_elnet}(object, metric = "mce", oneSE = TRUE,
  n_cores = 2, ...)

\method{summary}{beset_glm}(object, metric = NULL, n_pred = NULL,
  oneSE = TRUE, n_cores = 2, ...)

\method{summary}{beset_zeroinfl}(object, metric = "mce",
  n_count_pred = NULL, n_zero_pred = NULL, oneSE = TRUE, n_cores = 2,
  ...)
}
\arguments{
\item{object}{An object of class \code{"beset_glm"},
\code{"beset_zeroinfl"}, or \code{"beset_elnet"}}

\item{metric}{Prediction metric on which to based model selection. Can be one
of \code{"aic"} (Akaike information criterion), \code{"mae"} (mean absolute
error), \code{"mce"} (mean cross entropy), \code{"mse"} (mean squared error),
or \code{"r2"} (R-squared). If not specified, \code{"mse"} will be used for
Gaussian-family models and \code{"mce"} will be used for all other families.}

\item{oneSE}{Logical flag indicating whether or not the one-standard-error
rule should be applied when determining the best model from the model metric}

\item{n_cores}{Number of CPUs to use when computing a bootstrapped confidence
interval for the cross-validated estimate of R-squared}

\item{...}{additional arguments affecting the summary produced.}

\item{n_pred}{Optional: number of predictors that the best model should
contain}

\item{n_count_pred}{Optional: number of predictors that the count component
of a zero-inflation model should contain.}

\item{n_zero_pred}{Optional: number of predictors that the zero component of
a zero-inflation model should contain.}
}
\description{
Summary Methods for \code{beset} Objects
}
