% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_list_methods.R
\name{summary.cor_list}
\alias{summary.cor_list}
\title{Summarizing Correlation Lists}
\usage{
\method{summary}{cor_list}(object, x1 = everything(), x2 = everything(),
  sort = TRUE)
}
\arguments{
\item{object}{An object of class \code{\link{cor_list}}.}
}
\description{
Methods for the \code{cor_list} class.
}

