% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\alias{validate.lm}
\alias{validate.glm}
\alias{validate.glmnet}
\alias{validate.beset}
\alias{validate.nested}
\alias{validate.randomForest}
\alias{validate.beset_rf}
\title{Cross-Validated Prediction Metrics}
\usage{
validate(object, ...)

\method{validate}{lm}(object, data = NULL, n_folds = 10, n_reps = 10, seed = 42, ...)

\method{validate}{glm}(object, data = NULL, n_folds = 10, n_reps = 10, seed = 42, ...)

\method{validate}{glmnet}(
  object,
  x = NULL,
  y = NULL,
  lambda = NULL,
  offset = NULL,
  weights = NULL,
  n_folds = 10,
  n_reps = 10,
  seed = 42,
  envir = .GlobalEnv,
  ...
)

\method{validate}{beset}(object, ...)

\method{validate}{nested}(object, metric = "auto", oneSE = TRUE, ...)

\method{validate}{randomForest}(
  object,
  data = NULL,
  x = NULL,
  n_folds = 10,
  n_reps = 10,
  seed = 42,
  ...,
  parallel_type = NULL,
  n_cores = NULL,
  cl = NULL
)

\method{validate}{beset_rf}(object, ...)
}
\arguments{
\item{object}{A model object for which a cross-validated R-squared is
desired.}

\item{data}{Data frame that was used to train the model. Only needed if
the training data is not contained in the model \code{object}.}

\item{n_folds}{\code{Integer} indicating the number of folds to use for
cross-validation.}

\item{n_reps}{\code{Integer} indicating the number of times cross-validation
should be repeated.}

\item{seed}{\code{Integer} used to seed the random number generator.}

\item{x}{Model matrix that was used to train elastic net.}

\item{y}{Response variable that was used to train elastic net.}

\item{lambda}{\code{Numeric} value of the penalty parameter \code{lambda}
at which predictions are cross-validated. Default is the entire sequence
used to create the model, but importance scores will only be generated
if a single \code{lambda} value is given.}

\item{offset}{\code{Numeric} offset vector that was used to train elastic
net.}

\item{weights}{= \code{Numeric} offset vector that was used to train
elastic net.}

\item{envir}{\code{Environment} in which to look for variables listed in
object call}

\item{parallel_type}{(Optional) character string indicating the type of
parallel operation to be used, either \code{"fork"} or \code{"sock"}. If
omitted and \code{n_cores > 1}, the default is \code{"sock"} for Windows and
otherwise either \code{"fork"} or \code{"sock"} depending on which process is
being run.}

\item{n_cores}{Integer value indicating the number of workers to run in
parallel during subset search and cross-validation. By default, this will
be set to one fewer than the maximum number of physical cores you have
available, as indicated by \code{\link[parallel]{detectCores}}. Set to 1 to
disable parallel processing.}

\item{cl}{(Optional) \code{\link[parallel]{parallel}} or
\code{\link[snow]{snow}} cluster for use if \code{parallel_type = "sock"}.
If not supplied, a cluster on the local machine is automatically created.}
}
\value{
a "\code{cross_valid}" object consisting of a list with the
following elements:
\describe{
  \item{cv_stats}{a list of cross-validated prediction metrics, each
  containing the mean, between-fold standard error ("btwn_fold_se"), and
  between-repetition min-max range ("btwn_rep_range") of each metric.}
  \item{predictions}{a data frame containing the
  hold-out predictions for each row in the training data, with a separate
  column for each repetition of the k-fold cross-validation}
  \item{fold_assignments}{a data frame of equal
  dimensions to \code{predictions} giving the number of the hold-out fold
  of the corresponding element in \code{predictions}}
  \item{parameters}{a list documenting how many folds and repetitions were used
  for cross-validation, and the seed passed to the random number generator,
  which will be needed to reproduce the random fold assignments}
  }
}
\description{
\code{validate} is a generic function for cross-validating predictions from
the results of various model fitting functions. The function invokes
particular \code{\link[utils]{methods}} which depend on the
\code{\link[base]{class}} of the first argument.
}
\details{
To obtain cross-validation statistics, first fit a model as you
normally would and then pass the model object to \code{validate}.
}
\section{Methods (by class)}{
\itemize{
\item \code{lm}: Cross-validation of linear models

\item \code{glm}: Cross-validation of GLMs

\item \code{glmnet}: Cross-validation of GLM nets

\item \code{beset}: Cross-validation of beset objects

\item \code{nested}: Extract test error estimates from "beset_elnet"
objects with nested cross-validation

\item \code{randomForest}: Cross-validation of random forests

\item \code{beset_rf}: Retrieve cross-validation metrics from "beset_rf" object
}}

